chore: 🤖 Move Footer.tsx Header.tsx to shared
