+++ assets/react/components/shared/Footer.tsx
